chore(refactor): comments and moving a class
